Retrieving messages from Attrib::ExtAttrFact()
Summary:
There was no way to retrieve the error messages from items created
using Attrib::ExtAttrFact().create, as the create function was not returning
any object from which an error message could be read.
To fix it, ensured that all factory implementation have a name, and moved the
call to setTargetSelSpec to the external caller.

Test Plan:
Applied a volume processing setup which has an error in its IOPar on an inline,
the chain translator error message is now displayed in the GUI message.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D18196