Updated cmake test mechanism [devel]
Summary:
both 'run_test' csh and cmd scripts was incapable to handling arguments
which would have a space in the filepath.
This prevents for instance using CTest in an external OpendTect plugins repository, if OpendTect is installed in "C:\Program Files\OpendTect" for instance.
It proved impossible to specify a reliable machine command for Windows based on a Windows batch script (cmd), with spaces in the arguments, in combination with the cmake function add_test.
Instead, using a wrapper cmake script to safely handle all arguments, and launching the actual executable to test, instead of a script.
This new wrapper script needs to be included in the OpendTect SDK package.

Test Plan: All test programs run successfully.