Added rudimentary support for long-long literals [devel]
Summary:
The conversion from a long/long long type to a string should not trigger a compiler error.
This does not means that these types should be used in the code, they should now. But
adding a long/long object when creating a user message should not result in a compiler error, or
cross-platform compiler errors.

Test Plan:
Expanded test_string and test_basic, could not re-enabled some tests within test_threads.
Compiles, and runs on both Windows and Linux

Reviewers: nanne, khushnood

Reviewed By: khushnood

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D18230