Allow removing objects without switching IOMan::dirptr_
Summary: I have a use case where in the Translator::implRemove function of one object, I need to delete linked objects from another subdir of the survey. This fails because the IOMan::permRemove needs the dirptr_ to point to the subdir of the object being deleted, which is the first subdir. This change is a threadsafe way to achieve flexibility.

Test Plan: No adverse effect on existing functionality. Need to test the specific use case in the plugin.

Reviewers: arnaud, nanne

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D18231