Separating disp3d and disp2d for pick::set class
Summary:
Made Pick:Set::Disp class an interface class, while the actual implemenataions are now in Disp3D and Disp2D classes. All the old class members have been marked deprecated.

The polygon/polyline display properties( Linestyle and fill ) have been moved to a small class called PolyDisp. Disp has a pointer to PolyDisp and is only initialized if the set is a polygon. Although, the Disp class will initialize a PolyDisp object when asked for pointsets, it would do so with a pErrMsg.

Marker style properties have been moved to each subclass as MarkerStyle3D and MarkerStyle2D. The interface supports setting the style either via MarkerStyle3D/2D object or using the old way of type, size and color separately via setter functions. Although, the Disp class will initialize a PolyDisp object when asked for pointsets, it would do so with a pErrMsg.

As a result of this change, it is imperative that whenever a Pick::Set is created, it is always known before hand whether it is a pointset or a polygon.

Test Plan:
1) Displayed and created new polygons and pointsets in 3D scene and Basemap
2) Played around with PickSet and Polygons in the manager.
3) There are numerous pErrMsgs and Deprecated warnings. So, far didn't hit any in my tests.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18199