Initializing both 2d and 3d disp markers with same color for new picks
Summary: When a new polygon/pickset is created, the UI allows user to chose marker color. This color should apply to both 2d and 3d display properties. Also, depeneding on the object, line is either open or none and it should be the same for both 3d and 2d.

Test Plan: tested by picking new pointset and polygons in both 3D scene and Basemap.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18244