Fix rename for HDF5 wells
Summary: Renaming an HDF5 well did not rename the h5 file, thereby breaking the database. This was because a call to the base class function Translator::implRename was never made. While fixing it, I also reworked the archaic implementation of implRename and implRemove in odWellTranslator. The code is much simpler and easier to read now.

Test Plan: Renamed and removed both od and HDF5 wells successfully.

Reviewers: nanne

Differential Revision: https://phabricator.dgbes.com/D18256