Fixed Disappearing marker when editing of polygon in 3D scene
Summary:
The array containing marker's flag was resized, but the values were not updated accordingly. As a result, only the initial numbers of markers were visible at a time, depending on where the new point was inserted.

In this revision I am moving the values too for idxs after the insertion.

Test Plan: tested in 3D scene.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18259