Updated the ziphandler to support legacy zip packages
Summary:
While performing a installation of 7.0.0 using the new installer, the installer failed to unzip the downloaded packages.
The leagacy od application (7.0.0 or older) used the old zip tool to conpress and package the application.
But the new installer uses a new 7-zip tool for unzipping the packages during installation.
While debugging using the new installer, we analised that the "Created OS" information of Central directory header for 7.0.0  was read as "MVS or NTFS" for base_win64.zip.
Where NTFS is windows but MVS is IBM mainframe. Since the Created OS was a MVS, the filepath was falsly considered as directory and failed to recongnise it is a file.
To resolve the issues we also had to consider the "Extract OS" information to cross verify.

Test Plan:
An open source library(Zipdetails) was used to read the information on 7.0.0 and 8.0.0 packages for win, lux and macOS and list was populated to verify the problem and also to find the alternative.
After making the required changes, the packages were successfully unzipped for both 7.0.0 and 8.0.0.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D18263