'View Values' Set Transparency Bug Fix
Summary:
When you would go to view values, the segments would all be set to the
very first segment because the index would not iterate. That has been
fixed.

Test Plan: Compiled and ran ODT. When viewing segments, they showed up correctly in the 'View Points' and in the Color Table Manager Dialog.

Reviewers: raman, nanne

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com/D18268