Color Table Resets to Previous Transp Points When Going From 'Equal' to 'None' or 'Variable' Segmentation
Summary:
When you switch from 'Variable' or 'None' Segmentation to 'Equal' and
then back, the Color Table will look as it did before the initial
switch. That way if the user did not mean to switch it or wanted to keep
the initial values, they are able to do so.

Test Plan: Compiled and ran ODT. Switched back and forth between the different Color Table segmentations. Switched between the different tables in the list.

Reviewers: nanne, raman

Reviewed By: raman

Subscribers: arnaud, hadyan

Differential Revision: https://phabricator.dgbes.com/D18278