Support multi-component reading through SeisTrcReader::getDataPack
Summary: If the Translator has support for reading a DataPack directly, we bypass the Attribute engine and read sirectly through the SeisTrcReader::getDataPack function. This revision adds support for multi-component reading for this mode.

Test Plan: Tested reading a selected set of components from a multi-component dataset in a plugin implementation.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D18287