Updated OD::getSystemMemory implementation
Summary:

The Linux implementation was based on the old assumption that the available memory is the sum
of the free and cached memory. That is not correct, as not all cached memory is guaranteed to be
available for new allocations.
Since kernel 3.14, /proc/meminfo directly provides the 'available' memory, hence switching to this implementation,
already available in RHEL7.
Also removed the fetching of the swap/virtual memory, not used anywhere.
+ tentatively updated the macOS implementation for retrieving the total memory more directly.

Test Plan:

tested on Linux, consistent with the figures reported by various system monitoring tools.