Adding functionality to edit picksets/polygons
Summary:
Created a dialog to edit polygons or picksets.

polygon edit dialog has a table with all the locations and its co-ordinates that can be indivdually edited:
1) Values changed
2) Location removed
3) Bulk removal filtered on location of points ( Coord and Inl/Crl )

When this is launched from tree, it also has a button to edit z values in bulk to a Horizon or a constant Z.

Pointset edit dialog has a TextInfo with ranges information of the locations and spherical coordinates. Points cannot be indivdually edited as pointsets can be huge and thus can be impractical:
1) Bulk removal filtered on location of points ( Coord and Inl/Crl ) and its Sperical Coordinates( radius/dip/azimuth )

Test Plan: tested by editing polygons and pointsets

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18300