Updated to continue installation if there is a error in downloading the recommonded pkgs
Summary:
The process of installation is terminated if there is a python packge download failure. Since the python pkg is a recommonded package,
installation of this package can be done seperately and the process of configuration of od can be continued without the python pkg.
Also since the python pkg is a large file, donloading the pkg during poor network condition hinders'the process of OpendTect installation.
Jira: ODINST-49

Test Plan: Debugging using test case.

Reviewers: arnaud, khushnood

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D18289